Adiciona modularização no projeto
