--- src/main/java/com/projetoWeb/Arenas/controller/dto/CreateMatchDto.java
@@ -1,10 +1,0 @@
-package com.projetoWeb.Arenas.controller.dto;
-
-import java.time.ZonedDateTime;
-
-public record CreateMatchDto(ZonedDateTime matchData,
-                             String text,
-                             Long maxPlayers,
-                             String description,
-                             Long creatorUserId) {
-}
