--- src/main/java/com/projetoWeb/Arenas/controller/dto/CreateUserDto.java
@@ -1,16 +1,0 @@
-package com.projetoWeb.Arenas.controller.dto;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Size;
-
-public record CreateUserDto(
-        @NotBlank(message = "Email é obrigatório")
-        @Email(message = "Email deve ser válido")
-        String email,
-        
-        @NotBlank(message = "Senha é obrigatória")
-        @Size(min = 6, message = "Senha deve ter no mínimo 6 caracteres")
-        String password
-) {
-}
