--- src/main/java/com/projetoWeb/Arenas/controller/dto/LoginRequestDto.java
@@ -1,14 +1,0 @@
-package com.projetoWeb.Arenas.controller.dto;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-
-public record LoginRequestDto(
-        @NotBlank(message = "Email é obrigatório")
-        @Email(message = "Email deve ser válido")
-        String email,
-        
-        @NotBlank(message = "Senha é obrigatória")
-        String password
-) {
-}
