+++ src/main/java/com/projetoWeb/Arenas/controller/match/dto/CreateMatchDto.java
@@ -1,0 +1,10 @@
+package com.projetoWeb.Arenas.controller.match.dto;
+
+import java.time.ZonedDateTime;
+
+public record CreateMatchDto(ZonedDateTime matchData,
+    String text,
+    Long maxPlayers,
+    String description,
+    Long creatorUserId) {
+}
