+++ src/main/java/com/projetoWeb/Arenas/security/UserDetailsImpl.java
@@ -1,7 +1,8 @@
 package com.projetoWeb.Arenas.security;
 
-import com.projetoWeb.Arenas.service.UserService;
 import com.projetoWeb.Arenas.service.exception.EntityNotExistsException;
+import com.projetoWeb.Arenas.service.user.UserService;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
@@ -9,7 +10,7 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
 @Service
-public class UserDetailsImpl  implements UserDetailsService {
+public class UserDetailsImpl implements UserDetailsService {
 
     @Autowired
     private UserService userService;
